Jiggle RSpec/Cucumber versions to pick up some fixes we need.
rspec-mocks-2.6.0 includes a fix that allows an array containing mock objects to be flattened.

capybara-1.0.0 includes a fix to its use of the selenium-webdriver API, which was broken when using #select to select options from a select element.

Some of our Capybara domain juggling is no longer necessary with capybara-1.0.0.

The workaround for installing OrganisationResolvers under RSpec controller specs needed updated due to internal design changes in rspec-rails-2.6.0.