Don't let Rules template be deleted or get lost between test runs.
The 'truncation' strategy used by DatabaseCleaner (when running
Cucumber scenarios tagged '@javascript') was causing the Coop
rules template Document to be deleted after each scenario. It was
also deleting the Setting record which we used to find the correct
Document instance.

Now we tell DatabaseCleaner not to reset the Document and Paragraph
tables. We also have a fallback method for finding the correct
Document instance that doesn't use the Setting table.